perf: reduce the amount of work done when rendering data
Before, even if the limit of the number of render is set, it will render one more layer of
BaseLayout, which makes the performance can not be maximized, and now the optimization makes
BaseLayout will not render any more, even if the number of data is 1 million, it will only render
the specified amount of render. Performance has improved dramatically.

fix #352, fix #362, fix #258, fix #478