feat: replaced `onScrollBegin` with `onScrollStart`
The onScrollBegin event is triggered only when a finger touches down, which isn't consistent with
the onScrollEnd events. The correct approach would be to use the onScrollStart callback, ensuring
that both callbacks are invoked at the start and end of a scroll respectively.

BREAKING CHANGE: The onScrollBegin property has been deprecated. Please use onScrollStart as its
replacement.

fix #506